Move stories to component src
