+++ package.json
@@ -60,10 +60,12 @@
     "publish": "check-installed-dependencies && npm run compile && npm test && npm run props-table && lerna publish",
     "start": "cd packages/terra-consumer-site && npm run start",
     "start:express": "node scripts/express/app.js",
-    "storybook": "cd storybook && npm run storybook",
+    "storybook": "start-storybook -p 8081 -c .storybook",
     "test": "npm run jest && npm run nightwatch"
   },
   "devDependencies": {
+    "@storybook/addon-actions": "^3.2.14",
+    "@storybook/react": "^3.2.14",
     "babel-cli": "^6.24.1",
     "babel-core": "^6.24.1",
     "babel-jest": "^20.0.0",
@@ -73,6 +75,7 @@
     "babel-preset-react": "^6.24.1",
     "check-installed-dependencies": "^1.0.0",
     "chromedriver": "^2.31.0",
+    "css-loader": "^0.28.7",
     "danger": "^1.0.0",
     "enzyme": "^2.8.2",
     "enzyme-to-json": "^1.5.1",
@@ -88,19 +91,30 @@
     "lerna": "2.0.0-rc.4",
     "load-json-file": "^2.0.0",
     "nightwatch": "^0.9.16",
+    "node-sass": "^4.6.0",
+    "postcss-custom-properties": "^6.2.0",
+    "postcss-loader": "^2.0.8",
+    "postcss-rtl": "^1.1.3",
+    "prop-types": "^15.6.0",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
     "react-router": "^3.0.5",
     "react-router-dom": "^4.2.2",
     "react-test-renderer": "^15.5.4",
     "rimraf": "^2.6.1",
+    "sass-loader": "^6.0.6",
     "selenium-server-standalone-jar": "^3.4.0",
     "shelljs": "^0.7.7",
+    "style-loader": "^0.19.0",
     "stylelint": "~8.0.0",
     "stylelint-config-sass-guidelines": "^3.0.0",
     "stylelint-scss": "1.5.1",
     "stylelint-suitcss": "^1.0.0",
+    "terra-base": "^2.6.1",
     "terra-i18n": "^1.4.0",
+    "terra-i18n-plugin": "file:../terra-core/packages/terra-i18n-plugin",
+    "terra-icon": "^1.13.0",
+    "terra-table": "^1.15.0",
     "webpack": "^3.3.0",
     "webpack-dev-server": "^2.5.1"
   },
