+++ packages/terra-consumer-icon/stories/Index.jsx
