+++ packages/terra-consumer-layout/package.json
@@ -23,6 +23,7 @@
   },
   "homepage": "https://github.com/cerner/terra-consumer#readme",
   "peerDependencies": {
+    "@storybook/react": "^3.2.14",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-router-dom": "^4.2.2",
