+++ packages/terra-consumer-nav/package.json
@@ -23,6 +23,8 @@
   },
   "homepage": "https://github.com/cerner/terra-consumer#readme",
   "peerDependencies": {
+    "@storybook/addon-actions": "^3.2.14",
+    "@storybook/react": "^3.2.14",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "terra-base": "^2.1.0",
