--- storybook/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "terra-consumer-storybook",
-  "version": "1.0.0",
-  "description": "Doc site for terra-consumer",
-  "main": "config.js",
-  "author": "",
-  "license": "ISC",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/cerner/terra-consumer.git"
-  },
-  "scripts": {
-    "storybook": "start-storybook -p 8081 -c ."
-  },
-  "dependencies": {
-    "@storybook/addon-actions": "^3.2.13",
-    "@storybook/react": "^3.2.13",
-    "css-loader": "^0.28.7",
-    "node-sass": "^4.5.3",
-    "postcss-custom-properties": "^6.2.0",
-    "postcss-loader": "^2.0.8",
-    "postcss-rtl": "^1.1.3",
-    "prop-types": "^15.6.0",
-    "react": "^15.6.2",
-    "react-dom": "^15.6.2",
-    "react-router-dom": "^4.2.2",
-    "sass-loader": "^6.0.6",
-    "style-loader": "^0.19.0",
-    "terra-base": "^2.6.1",
-    "terra-consumer-icon": "file:../packages/terra-consumer-icon",
-    "terra-consumer-layout": "file:../packages/terra-consumer-layout",
-    "terra-consumer-nav": "file:../packages/terra-consumer-nav",
-    "terra-i18n": "^1.10.1",
-    "terra-i18n-plugin": "^1.8.0",
-    "terra-icon": "^1.12.1",
-    "terra-table": "^1.14.0"
-  }
-}
